Added decimal

+++ lab1/modules/series.py
@@ -1,9 +1,10 @@
 from __future__ import division
+from decimal import Decimal
 
 # Series module
 
 def sumOfSeries(n):
-    sum = 0.0
+    sum = Decimal(0.0)
     for i in range(1, n+1):
-        sum += pow(i / (i + 1), pow(i, 2))
-    return sum
+        sum += Decimal(pow(i / (i + 1), pow(i, 2)))
+    return Decimal(sum)
